+++ lib/generators/marionette/layout/templates/app/templates/item_view.jst.eco
@@ -1,0 +1,1 @@
+<p>Find me in app/templates/<%= title.underscore %></p>
