+++ lib/generators/marionette/layout/templates/app/views/item_view.js.coffee
@@ -1,0 +1,4 @@
+@Backbone.app.module "Views", (Views, App, Backbone, Marionette, $, _) ->
+
+  class Views.<%= title.capitalize %>Layout extends App.Views.ItemView
+    template: '<%= title.underscore %>'
