+++ lib/generators/marionette/layout/templates/app/views/layouts/layouts.js.coffee
@@ -1,6 +1,6 @@
 @Backbone.app.module "Views", (Views, App, Backbone, Marionette, $, _) ->
 
   class Views.<%= title.capitalize %>Layout extends App.Views.Layout
-    template: 'layouts/<%= title %>'
+    template: 'layouts/<%= title.underscore %>'
     regions:
       bodyRegion: "#body"
